backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
pydantic==2.5.3
librosa==0.10.1
numpy==1.26.3
web3==6.14.0
python-dotenv==1.0.0
aiofiles==23.2.1
